# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: AI Quant Research Assistant API
  version: 1.0.0
  summary: Asynchronous, evidence-backed US equity research API
  description: |
    Public contract for the Java orchestration service. All financial facts must be
    traceable to registered evidence. Research created with mock providers is
    marked as MOCK; the UI and exports render DEMO DATA - NOT REAL MARKET DATA.
    MIXED_TEST is test-only and cannot be published through ordinary user flows.
    During Phase 2, Basic authentication is available only through an explicitly
    enabled development/test demo principal. Bearer authentication is the production
    contract but is not activated until the formal auth integration is delivered;
    the Phase 2 production profile therefore fails closed at startup.
  contact:
    name: AI Quant Research Assistant maintainers
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
  - basicDemoAuth: []
paths:
  /api/v1/research:
    post:
      tags: [Research]
      operationId: createResearch
      summary: Create and enqueue a research job
      description: |
        Locally known inactive/unsupported securities return INVALID_SYMBOL, and a
        locally provable symbol/company-name contradiction returns SECURITY_MISMATCH.
        Unknown but well-formed securities are queued for durable provider-backed
        RESOLVE_SECURITY rather than resolved through a request-thread network call.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResearchRequest'
            examples:
              muResearch:
                value:
                  query: 分析 MU 未来 12 个月的主要增长动力和风险
                  symbol: MU
                  locale: zh-CN
                  benchmark: QQQ
                  period: 5y
                  reportDepth: STANDARD
                  includeTechnicalAnalysis: true
                  includeFundamentalAnalysis: true
                  includeMacroAnalysis: true
      responses:
        '202':
          description: Research job accepted and queued
          headers:
            Location:
              description: Canonical URL of the created research job
              schema:
                type: string
            Retry-After:
              description: Suggested status polling delay in seconds
              schema:
                type: integer
                minimum: 1
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags: [Research]
      operationId: listResearch
      summary: List visible research jobs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: symbol
          in: query
          description: Exact, case-insensitive ticker symbol
          schema:
            type: string
            pattern: '^[A-Za-z][A-Za-z0-9.-]{0,9}$'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ResearchStatus'
        - name: from
          in: query
          description: Include jobs created on or after this UTC instant
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Include jobs created before this UTC instant
          schema:
            type: string
            format: date-time
        - name: q
          in: query
          description: Search title, query, symbol, or company name
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - createdAt,desc
              - createdAt,asc
              - updatedAt,desc
              - updatedAt,asc
            default: createdAt,desc
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Paginated research jobs; soft-deleted jobs are excluded
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/research/{researchId}:
    parameters:
      - $ref: '#/components/parameters/ResearchId'
      - $ref: '#/components/parameters/XRequestId'
    get:
      tags: [Research]
      operationId: getResearch
      summary: Get research details and the latest report, if available
      responses:
        '200':
          description: Research details
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Research]
      operationId: deleteResearch
      summary: Soft-delete a research job
      description: |
        Hides the research job from normal reads. Audit records, source snapshots,
        evidence and report versions are retained according to the retention policy.
        Repeated deletion is successful and returns 204.
      responses:
        '204':
          description: Research job is now soft-deleted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/research/{researchId}/status:
    parameters:
      - $ref: '#/components/parameters/ResearchId'
      - $ref: '#/components/parameters/XRequestId'
    get:
      tags: [Research]
      operationId: getResearchStatus
      summary: Get research progress and step attempts
      responses:
        '200':
          description: Current status snapshot
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/research/{researchId}/retry:
    parameters:
      - $ref: '#/components/parameters/ResearchId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/XRequestId'
    post:
      tags: [Research]
      operationId: retryResearch
      summary: Resume a failed or partially completed research job
      description: Successful steps with unchanged input hashes are reused.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryResearchRequest'
      responses:
        '202':
          description: Retry accepted
          headers:
            Location:
              description: Status URL for this research job
              schema:
                type: string
            Retry-After:
              schema:
                type: integer
                minimum: 1
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/research/{researchId}/cancel:
    parameters:
      - $ref: '#/components/parameters/ResearchId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/XRequestId'
    post:
      tags: [Research]
      operationId: cancelResearch
      summary: Request cooperative cancellation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelResearchRequest'
      responses:
        '202':
          description: |
            Cancellation request accepted. COMPLETED, PARTIALLY_COMPLETED, and
            CANCELLED are accepted no-ops returning the existing snapshot; FAILED
            returns 409 because retry/delete, not cancellation, is the valid action.
          headers:
            Location:
              description: Status URL for this research job
              schema:
                type: string
            Retry-After:
              schema:
                type: integer
                minimum: 1
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production authentication contract. Its implementation is intentionally
        deferred beyond Phase 2; Phase 2 production startup fails closed.
    basicDemoAuth:
      type: http
      scheme: basic
      description: |
        Explicit development/test demo principal only. It is rejected in production.
  parameters:
    ResearchId:
      name: researchId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Unique key scoped to authenticated user, method, and path. Reusing a key
        with a different request body returns IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 100
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page number
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  headers:
    XRequestId:
      description: Request correlation identifier
      schema:
        type: string
    IdempotencyReplayed:
      description: True when the response was replayed from an earlier request
      schema:
        type: boolean
  responses:
    BadRequest:
      description: Request syntax or parameter validation failed
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication is missing or invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Caller cannot access this resource or the account is not active
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Route/resource does not exist, is inaccessible, or is soft-deleted
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Current state conflicts with the requested operation
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: Request is syntactically valid but semantically invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: Local or upstream rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until retry is advised
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServiceUnavailable:
      description: A required service is temporarily unavailable
      headers:
        Retry-After:
          description: Seconds until retry is advised, when known
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ResearchStatus:
      type: string
      enum:
        - CREATED
        - QUEUED
        - RESOLVING_SECURITY
        - FETCHING_MARKET_DATA
        - FETCHING_FUNDAMENTALS
        - FETCHING_FILINGS
        - FETCHING_MACRO_DATA
        - VALIDATING_DATA
        - RUNNING_QUANT_ANALYSIS
        - ANALYZING_FUNDAMENTALS
        - BUILDING_EVIDENCE
        - GENERATING_REPORT
        - VALIDATING_REPORT
        - COMPLETED
        - PARTIALLY_COMPLETED
        - FAILED
        - CANCELLED
    StepType:
      type: string
      enum:
        - RESOLVE_SECURITY
        - FETCH_MARKET_DATA
        - FETCH_FUNDAMENTALS
        - FETCH_FILINGS
        - FETCH_MACRO_DATA
        - VALIDATE_DATA
        - RUN_QUANT_ANALYSIS
        - ANALYZE_FUNDAMENTALS
        - BUILD_EVIDENCE
        - GENERATE_REPORT
        - VALIDATE_REPORT
...
    ReportDepth:
      type: string
      enum: [QUICK, STANDARD, DEEP]
      default: STANDARD
    ResearchPeriod:
      type: string
      enum: [1y, 3y, 5y, 10y, max]
      default: 5y
    Locale:
      type: string
      enum: [zh-CN, en-US]
    DataMode:
      type: string
      description: |
        REAL uses only real providers. MOCK uses only deterministic mock data and
        must display DEMO DATA - NOT REAL MARKET DATA in the UI and every export.
        MIXED_TEST may mix real and mock inputs only in automated tests or fault
        drills and must never be published through ordinary user flows.
      enum: [REAL, MOCK, MIXED_TEST]
...
    CreateResearchRequest:
      type: object
      additionalProperties: false
      required:
        - query
      anyOf:
        - required: [symbol]
        - required: [companyName]
      properties:
        query:
          type: string
          minLength: 10
          maxLength: 4000
        symbol:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9.-]{0,9}$'
        companyName:
          type: string
          minLength: 1
          maxLength: 200
        locale:
          allOf:
            - $ref: '#/components/schemas/Locale'
          description: Inferred from the query language when omitted.
        benchmark:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9.-]{0,9}$'
          default: SPY
        period:
          $ref: '#/components/schemas/ResearchPeriod'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reportDepth:
          $ref: '#/components/schemas/ReportDepth'
        includeTechnicalAnalysis:
          type: boolean
          default: true
        includeFundamentalAnalysis:
          type: boolean
          default: true
        includeMacroAnalysis:
          type: boolean
          default: true
      description: |
        Either period or startDate/endDate may be used. If both are supplied,
        explicit dates take precedence after validation.
    RetryResearchRequest:
      type: object
      additionalProperties: false
      properties:
        fromStep:
          $ref: '#/components/schemas/StepType'
        reason:
          type: string
          maxLength: 500
    CancelResearchRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 500
    ResearchLinks:
      type: object
      additionalProperties: false
      required: [self, status]
      properties:
        self:
          type: string
        status:
          type: string
        evidence:
          type: string
        reports:
          type: string
        export:
          type: string
    ResearchAccepted:
      type: object
      additionalProperties: false
      required: [researchId, status, dataMode, createdAt, links]
      properties:
        researchId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ResearchStatus'
        dataMode:
          $ref: '#/components/schemas/DataMode'
        createdAt:
          type: string
          format: date-time
        links:
          $ref: '#/components/schemas/ResearchLinks'
    ResearchItem:
      type: object
      required:
        - researchId
        - query
        - symbol
        - companyName
        - status
        - progress
        - reportDepth
        - dataMode
        - createdAt
        - updatedAt
      properties:
        researchId:
          type: string
          format: uuid
        title:
          type: [string, 'null']
        query:
          type: string
        symbol:
          type: [string, 'null']
        companyName:
          type: [string, 'null']
        benchmark:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/ResearchStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        reportDepth:
          $ref: '#/components/schemas/ReportDepth'
        dataMode:
          $ref: '#/components/schemas/DataMode'
        latestReportVersion:
          type: [integer, 'null']
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
    ResearchDetail:
      allOf:
        - $ref: '#/components/schemas/ResearchItem'
        - type: object
          required:
            - request
            - currentStep
            - cancellationRequested
            - warnings
            - links
          properties:
            request:
              $ref: '#/components/schemas/CreateResearchRequest'
            currentStep:
              type: [string, 'null']
            startedAt:
              type: [string, 'null']
              format: date-time
            cancellationRequested:
              type: boolean
            lastError:
              oneOf:
                - $ref: '#/components/schemas/ErrorSummary'
                - type: 'null'
            latestReport:
              oneOf:
                - $ref: '#/components/schemas/ReportVersion'
                - type: 'null'
            warnings:
              type: array
              items:
                $ref: '#/components/schemas/Warning'
            links:
              $ref: '#/components/schemas/ResearchLinks'
      unevaluatedProperties: false
...
    ResearchStatusResponse:
      type: object
      additionalProperties: false
      required:
        - researchId
        - status
        - progress
        - completedSteps
        - totalSteps
        - cancellationRequested
        - dataMode
        - steps
        - updatedAt
      properties:
        researchId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ResearchStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        currentStep:
          type: [string, 'null']
        completedSteps:
          type: integer
          minimum: 0
        totalSteps:
          type: integer
          minimum: 1
        cancellationRequested:
          type: boolean
        dataMode:
          $ref: '#/components/schemas/DataMode'
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorSummary'
            - type: 'null'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ResearchStep'
        updatedAt:
          type: string
          format: date-time
    PageMetadata:
      type: object
      additionalProperties: false
      required: [number, size, totalElements, totalPages, first, last]
      properties:
        number:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    ResearchPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResearchItem'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    ReportVersion:
      allOf:
        - $ref: '#/components/schemas/ReportVersionSummary'
        - type: object
          required:
            - reportSchemaVersion
            - generatedAt
            - executiveSummary
            - claims
            - sections
            - evidenceIds
            - dataQuality
            - warnings
            - disclaimer
          properties:
            reportSchemaVersion:
              type: string
              const: research_report_schema_v1
            generatedAt:
              type: string
              format: date-time
            executiveSummary:
              type: string
            claims:
              type: array
              items:
                $ref: '#/components/schemas/Claim'
            sections:
              type: object
              description: Versioned structured report sections; never unvalidated free-form JSON.
              additionalProperties: true
            evidenceIds:
              type: array
              uniqueItems: true
              items:
                type: string
                pattern: '^ev_[A-Za-z0-9]+$'
            dataQuality:
              $ref: '#/components/schemas/DataQuality'
            warnings:
              type: array
              items:
                $ref: '#/components/schemas/Warning'
            disclaimer:
              type: string
      unevaluatedProperties: false
...
    Warning:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        evidenceIds:
          type: array
          uniqueItems: true
          items:
            type: string
            pattern: '^ev_[A-Za-z0-9]+$'
    ErrorSummary:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        failedStep:
          oneOf:
            - $ref: '#/components/schemas/StepType'
            - type: 'null'
    ApiError:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - timestamp
        - status
        - code
        - message
        - requestId
        - researchId
        - retryable
        - details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        message:
          type: string
        requestId:
          type: string
        researchId:
          type: [string, 'null']
          format: uuid
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true